requirements.txt
fastapi
uvicorn
openai
pydantic
pillow
chromadb
sentence-transformers
transformers
accelerate
safetensors
torch
python-dotenv
peft
